requirements.txt
praw
pyjwt
